+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix W to place prefab not working on prefabs placed in previous sessions
 - Significantly speed up auto transparency detection
 - Fix temporary error when upgrading version from 1.x -> 1.8.x
+- Increased clarity on material mapping menu so conflicts are less likely to be ignored
 
 # 1.8.1
 - Fix scaling for linked prefabs